fix(vcs_git): drop the `--graph' parameter for `gl'
The `gl' alias is used to display a one-line-style Git log. However, the
original alias comes with the `--graph' parameter, which hugely slows down the
command response time, especially in larger repositories such as the Linux
Kernel and Glibc, etc.

Since the graphing feature (for illustrating branch relationships) is rarely
used in this scenario, drop this parameter in favor of better performance.